flytt funksjon for oppdatering av andeler ut i egen generator-fil + flytt relevante tester
